refactor: extract FramePlayerBase ABC and rename FrameCapture to BlackHoleFramePlayer
Introduce a shared FramePlayerBase interface for frame players, move
the capture implementation to its own file as BlackHoleFramePlayer,
and have LedFramePlayer inherit from the same ABC.
