+++ pifi/led/ledframeplayer.py
@@ -4,12 +4,13 @@ import numpy as np
 
 from pifi.config import Config
 from pifi.directoryutils import DirectoryUtils
+from pifi.led.frameplayerbase import FramePlayerBase
 from pifi.led.gamma import Gamma
 from pifi.led.drivers.leddrivers import LedDrivers
 from pifi.settingsdb import SettingsDb
 from pifi.video.videocolormode import VideoColorMode
 
-class LedFramePlayer:
+class LedFramePlayer(FramePlayerBase):
 
     __FADE_STEPS = 5
